Added abstractions for vertex buffers, index buffers, vertex arrays, shaders, and improved some of the other code. Also cleaned up the folder hierarchy because having both an include and src with subdirectories was getting cluttered

+++ Ruby/src/PlatformSpecific/Windows/WindowsWindow.cpp
@@ -109,7 +109,7 @@ namespace Ruby {
 
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xoffset, double yoffset) {
 				
-				MouseScrollEvent e(xoffset, yoffset);
+				MouseScrollEvent e((float)xoffset, (float)yoffset);
 				App::getInstance().onEvent(e);
 
 			});
